requirements.txt
streamlit>=1.33.0
pandas>=2.0.0
requests>=2.28.0
python-dateutil>=2.8.2
altair>=5.0.0